♻️ :: (#29) move entities into domain package
