+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/picnicapply/domain/PicnicApply.java
@@ -1,4 +1,4 @@
-package io.github.v1serviceapplication.domain.picnicapply;
+package io.github.v1serviceapplication.domain.picnicapply.domain;
 
 import io.github.v1serviceapplication.global.entity.BaseUUIDEntity;
 import lombok.*;
