+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stayapply/domain/StayApply.java
@@ -1,4 +1,4 @@
-package io.github.v1serviceapplication.domain.stayapply;
+package io.github.v1serviceapplication.domain.stayapply.domain;
 
 import io.github.v1serviceapplication.global.entity.BaseUUIDEntity;
 import lombok.*;
